+++ tsl/src/nodes/vector_agg/hashing/hash_strategy_impl_single_fixed_key.c
@@ -54,14 +54,14 @@ FUNCTION_NAME(key_hashing_get_key)(BatchHashingParams params, int row,
 }
 
 static pg_attribute_always_inline void
-FUNCTION_NAME(key_hashing_store_new)(HashingStrategy *restrict hashing, uint32 new_key_index,
+FUNCTION_NAME(key_hashing_store_new)(HashingStrategy *restrict hashing, uint16 new_key_index,
 									 OUTPUT_KEY_TYPE output_key)
 {
 	hashing->output_keys[new_key_index] = OUTPUT_KEY_TO_DATUM(output_key);
 }
 
 static void
-FUNCTION_NAME(emit_key)(GroupingPolicyHash *policy, uint32 current_key,
+FUNCTION_NAME(emit_key)(GroupingPolicyHash *policy, uint16 current_key,
 						TupleTableSlot *aggregated_slot)
 {
 	hash_strategy_output_key_single_emit(policy, current_key, aggregated_slot);
